Final project with simplified html
I eliminated player names and just gave them numbers.
I eliminated all of the unnecessary columns.
Under Round 1, I replaced one content editable td with a td with a form
element with an input tag and submit button on lines 114-119. I left
the action as "" since I don't know what I should put there. I gave it
a bg color of red just to highlight it for our puposes.
On the Scoring tab, I gave the td on line 87 a bg color of red also
just to highlight it for our purposes.
Here is the Big Question. Can the score that is entered by the user and
submitted (lines 114-119) be returned to the scoring page td (line 87)?