Simplified html for final project
For the time being, I have eliminated player names. All players are
identified by numbers.
I eliminated the unnecessary columns that would be needed only for a
more complicated scoring method.
In Round 1, I replaced one content editable td with a td containing a
form element with input tag and submit button. I gave it a bg color of
red just to highlight it for our purposes. I left the action as ""
because I did not know what to put there. The lines are 114-119.
The Big Question is whether the score that is entered by the user in
the Round 1 td (lines 114-119) can be returned to the appropriate td
(line 87) on the Scoring tab.